adicionando nova nomenclatura
